magit-ediff-quit-hook: new hook variable
We set the local value of `ediff-quit-hook' to a function specific
to the Ediff variant being used.  This has the advantage, that
`ediff-cleanup-mess' is not run; that function is not suitable for
our needs.  But it also means that additional functions added by
the user will be ignored.

So add an alternative hook variable `magit-ediff-quit-hook', which
replaces `ediff-quit-hook' when Ediff is invoked using a Magit
command.

Add the existing function `magit-ediff-cleanup-auxiliary-buffer' and
the new function `magit-ediff-restore-previous-winconf' to the new
hook variable.  Users who add something to `magit-ediff-quit-hook'
will likely want to remove `magit-ediff-restore-previos-winconf'.